v 17.9 Refactoring gathering services
